+++ Tests/SentryTests/SentryHubTests.swift
@@ -22,6 +22,7 @@ class SentryHubTests: XCTestCase {
         let transactionName = "Some Transaction"
         let transactionOperation = "Some Operation"
         let random = TestRandom(value: 0.5)
+        let queue = DispatchQueue(label: "SentryHubTests", qos: .utility, attributes: [.concurrent])
         
         init() {
             options = Options()
@@ -383,12 +384,29 @@ class SentryHubTests: XCTestCase {
         if let errorArguments = fixture.client.captureErrorWithSessionInvocations.first {
             XCTAssertEqual(fixture.error, errorArguments.error as NSError)
             
-            XCTAssertEqual(1, errorArguments.session.errors)
-            XCTAssertEqual(SentrySessionStatus.ok, errorArguments.session.status)
+            XCTAssertEqual(1, errorArguments.session?.errors)
+            XCTAssertEqual(SentrySessionStatus.ok, errorArguments.session?.status)
             
             XCTAssertEqual(fixture.scope, errorArguments.scope)
         }
         
+        // only session init is sent
+        XCTAssertEqual(1, fixture.client.captureSessionInvocations.count)
+    }
+
+    func testCaptureWithoutIncreasingErrorCount() {
+        let sut = fixture.getSut()
+        sut.startSession()
+        fixture.client.callSessionBlockWithIncrementSessionErrors = false
+        sut.capture(error: fixture.error, scope: fixture.scope).assertIsNotEmpty()
+
+        XCTAssertEqual(1, fixture.client.captureErrorWithSessionInvocations.count)
+        if let errorArguments = fixture.client.captureErrorWithSessionInvocations.first {
+            XCTAssertEqual(fixture.error, errorArguments.error as NSError)
+            XCTAssertNil(errorArguments.session)
+            XCTAssertEqual(fixture.scope, errorArguments.scope)
+        }
+
         // only session init is sent
         XCTAssertEqual(1, fixture.client.captureSessionInvocations.count)
     }
@@ -432,12 +450,29 @@ class SentryHubTests: XCTestCase {
         if let exceptionArguments = fixture.client.captureExceptionWithSessionInvocations.first {
             XCTAssertEqual(fixture.exception, exceptionArguments.exception)
             
-            XCTAssertEqual(1, exceptionArguments.session.errors)
-            XCTAssertEqual(SentrySessionStatus.ok, exceptionArguments.session.status)
+            XCTAssertEqual(1, exceptionArguments.session?.errors)
+            XCTAssertEqual(SentrySessionStatus.ok, exceptionArguments.session?.status)
             
             XCTAssertEqual(fixture.scope, exceptionArguments.scope)
         }
         
+        // only session init is sent
+        XCTAssertEqual(1, fixture.client.captureSessionInvocations.count)
+    }
+
+    func testCaptureExceptionWithoutIncreasingErrorCount() {
+        let sut = fixture.getSut()
+        sut.startSession()
+        fixture.client.callSessionBlockWithIncrementSessionErrors = false
+        sut.capture(exception: fixture.exception, scope: fixture.scope).assertIsNotEmpty()
+
+        XCTAssertEqual(1, fixture.client.captureExceptionWithSessionInvocations.count)
+        if let exceptionArguments = fixture.client.captureExceptionWithSessionInvocations.first {
+            XCTAssertEqual(fixture.exception, exceptionArguments.exception)
+            XCTAssertNil(exceptionArguments.session)
+            XCTAssertEqual(fixture.scope, exceptionArguments.scope)
+        }
+
         // only session init is sent
         XCTAssertEqual(1, fixture.client.captureSessionInvocations.count)
     }
@@ -456,7 +491,7 @@ class SentryHubTests: XCTestCase {
         for i in 1...captureCount {
             // The session error count must not be in order as we use a concurrent DispatchQueue
             XCTAssertTrue(
-                invocations.contains { $0.session.errors == i },
+                invocations.contains { $0.session!.errors == i },
                 "No session captured with \(i) amount of errors."
             )
         }
@@ -476,7 +511,7 @@ class SentryHubTests: XCTestCase {
         for i in 1..<captureCount {
             // The session error count must not be in order as we use a concurrent DispatchQueue
             XCTAssertTrue(
-                invocations.contains { $0.session.errors == i },
+                invocations.contains { $0.session!.errors == i },
                 "No session captured with \(i) amount of errors."
             )
         }
@@ -639,8 +674,7 @@ class SentryHubTests: XCTestCase {
         let sut = fixture.getSut()
         sut.startSession()
 
-        let queue = DispatchQueue(label: "SentryHubTests", qos: .utility, attributes: [.concurrent])
-
+        let queue = fixture.queue
         let group = DispatchGroup()
         for _ in 0..<count {
             group.enter()
@@ -651,6 +685,69 @@ class SentryHubTests: XCTestCase {
         }
 
         group.waitWithTimeout()
+    }
+    
+    @available(iOS 10.0, tvOS 10.0, OSX 10.12, *)
+    func testModifyIntegrationsConcurrently() {
+        
+        let sut = fixture.getSut()
+        
+        let outerLoopAmount = 10
+        let innerLoopAmount = 100
+        
+        let queue = fixture.queue
+        let group = DispatchGroup()
+    
+        for i in 0..<outerLoopAmount {
+            group.enter()
+            queue.async {
+                for j in 0..<innerLoopAmount {
+                    let integrationName = "Integration\(i)\(j)"
+                    sut.addInstalledIntegration(EmptyIntegration(), name: integrationName)
+                    XCTAssertTrue(sut.hasIntegration(integrationName))
+                }
+                group.leave()
+            }
+        }
+        
+        group.waitWithTimeout()
+        
+        XCTAssertEqual(innerLoopAmount * outerLoopAmount, sut.installedIntegrations.count)
+        XCTAssertEqual(innerLoopAmount * outerLoopAmount, sut.installedIntegrationNames.count)
+        
+    }
+    
+    /**
+     * This test only ensures concurrent modifications don't crash.
+     */
+    @available(iOS 10.0, tvOS 10.0, OSX 10.12, *)
+    func testModifyIntegrationsConcurrently_NoCrash() {
+        let sut = fixture.getSut()
+        
+        let queue = fixture.queue
+        let group = DispatchGroup()
+        
+        for i in 0..<1_000 {
+            group.enter()
+            queue.async {
+                for j in 0..<10 {
+                    let integrationName = "Integration\(i)\(j)"
+                    sut.addInstalledIntegration(EmptyIntegration(), name: integrationName)
+                    sut.hasIntegration(integrationName)
+                    sut.isIntegrationInstalled(EmptyIntegration.self)
+                }
+                XCTAssertLessThanOrEqual(0, sut.installedIntegrations.count)
+                sut.installedIntegrations.forEach { XCTAssertNotNil($0) }
+                
+                XCTAssertLessThanOrEqual(0, sut.installedIntegrationNames.count)
+                sut.installedIntegrationNames.forEach { XCTAssertNotNil($0) }
+                sut.removeAllIntegrations()
+                
+                group.leave()
+            }
+        }
+        
+        group.wait()
     }
     
     private func captureEventEnvelope(level: SentryLevel) {